+++ slang/bin/clean.dart
@@ -1,0 +1,5 @@
+import 'slang.dart' as mainRunner;
+
+void main(List<String> arguments) async {
+  mainRunner.main(['clean', ...arguments]);
+}